+++ src/query/unix/ram.cpp
@@ -90,20 +90,20 @@ RAM::RAM() noexcept
 }
 
 // clang-format off
-float& RAM::free_amount() noexcept
+double& RAM::free_amount() noexcept
 { return m_memory_infos.free_amount; }
 
-float& RAM::total_amount() noexcept
+double& RAM::total_amount() noexcept
 { return m_memory_infos.total_amount; }
 
-float& RAM::used_amount() noexcept
+double& RAM::used_amount() noexcept
 { return m_memory_infos.used_amount; }
 
-float& RAM::swap_total_amount() noexcept
+double& RAM::swap_total_amount() noexcept
 { return m_memory_infos.swap_total_amount; }
 
-float& RAM::swap_used_amount() noexcept
+double& RAM::swap_used_amount() noexcept
 { return m_memory_infos.swap_used_amount; }
 
-float& RAM::swap_free_amount() noexcept
+double& RAM::swap_free_amount() noexcept
 { return m_memory_infos.swap_free_amount; }
